# Optimize RS Controller Performance: Reduce Work Duration Time & Minimize Cache Locking