@soltysh 2025-05-07
> /assign @soltysh

Sorry, this slipped through the cracks. @hakuna-matatah can you rebase the PR, fix that linter error (adding `//nolint:errcheck` is ok for that one the linter is failing on, and then ping me on slack for a re-review on this one?